# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subcategory {
  id: number;
  categoryId: number;
  name: string;
  slug: string;
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  createdAt: string;
  subcategories?: Subcategory[];
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  image?: string;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  image?: string;
}

export interface SubcategoryInput {
  name: string;
  slug: string;
}

export interface SubcategoryUpdate {
  name?: string;
  slug?: string;
}

export interface ProductSpec {
  /** @nullable */
  capacity?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  discountPct?: number | null;
  /** @nullable */
  discountedPrice?: number | null;
  images: string[];
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  category?: Category | null;
  subcategory?: Subcategory | null;
  stockQty?: number;
  inStock: boolean;
  visible: boolean;
  featured: boolean;
  newArrival: boolean;
  bestSeller: boolean;
  specs?: ProductSpec;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  discountPct?: number;
  images: string[];
  coverImage?: string;
  categoryId?: number;
  subcategoryId?: number;
  stockQty?: number;
  inStock?: boolean;
  visible?: boolean;
  featured?: boolean;
  newArrival?: boolean;
  bestSeller?: boolean;
  specs?: ProductSpec;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  discountPct?: number | null;
  images?: string[];
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  stockQty?: number;
  inStock?: boolean;
  visible?: boolean;
  featured?: boolean;
  newArrival?: boolean;
  bestSeller?: boolean;
  specs?: ProductSpec;
}

export interface ProductsSummary {
  total: number;
  visible: number;
  hidden: number;
  featured: number;
  newArrivals: number;
  bestSellers: number;
  outOfStock: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productImage: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface OrderTimeline {
  status: string;
  timestamp: string;
  /** @nullable */
  note?: string | null;
}

export interface Order {
  id: number;
  orderRef: string;
  fullName: string;
  phone: string;
  state: string;
  city: string;
  address: string;
  items: OrderItem[];
  subtotal: number;
  paymentMethod: string;
  paymentStatus: string;
  orderStatus: string;
  /** @nullable */
  flutterwaveRef?: string | null;
  timeline: OrderTimeline[];
  createdAt: string;
}

export interface OrderInput {
  fullName: string;
  phone: string;
  state: string;
  city: string;
  address: string;
  paymentMethod: string;
  flutterwaveRef?: string;
  items: OrderItem[];
}

export interface OrderStatusUpdate {
  orderStatus: string;
  paymentStatus?: string;
  note?: string;
}

export interface OrderStats {
  totalOrders: number;
  pendingOrders: number;
  paidOrders: number;
  preparingOrders: number;
  deliveredOrders?: number;
  cancelledOrders?: number;
  totalRevenue: number;
}

export interface TrustItem {
  id: string;
  title: string;
  subtitle: string;
  icon: string;
}

export interface HeroBanner {
  id: string;
  title: string;
  subtitle: string;
  image: string;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  buttonLink?: string | null;
}

export interface FeaturedCollection {
  id: string;
  title: string;
  description: string;
  image: string;
  link: string;
}

export interface HomepageSettings {
  heroBanners: HeroBanner[];
  trustItems: TrustItem[];
  featuredCollections: FeaturedCollection[];
}

export interface HomepageSettingsUpdate {
  heroBanners?: HeroBanner[];
  trustItems?: TrustItem[];
  featuredCollections?: FeaturedCollection[];
}

export interface SocialLinks {
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  tiktok?: string | null;
}

export interface DeliverySettings {
  /** @nullable */
  freeDeliveryThreshold?: number | null;
  deliveryFee?: number;
  /** @nullable */
  deliveryNote?: string | null;
}

export interface StoreSettings {
  storeName: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  email: string;
  socialLinks?: SocialLinks;
  deliverySettings?: DeliverySettings;
}

export interface StoreSettingsUpdate {
  storeName?: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  email?: string;
  socialLinks?: SocialLinks;
  deliverySettings?: DeliverySettings;
}

export interface PaymentSettings {
  flutterwaveEnabled: boolean;
  payOnDeliveryEnabled: boolean;
  /** @nullable */
  flutterwavePublicKey?: string | null;
}

export interface PaymentSettingsUpdate {
  flutterwaveEnabled?: boolean;
  payOnDeliveryEnabled?: boolean;
}

export type ListProductsParams = {
categoryId?: number;
subcategoryId?: number;
featured?: boolean;
newArrival?: boolean;
bestSeller?: boolean;
search?: string;
visible?: boolean;
};

export type ListOrdersParams = {
status?: string;
};