# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MimiiHub e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with subcategories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: Deleted
  /categories/{categoryId}/subcategories:
    post:
      operationId: createSubcategory
      tags: [categories]
      summary: Create a subcategory
      parameters:
        - { name: categoryId, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
  /subcategories/{id}:
    patch:
      operationId: updateSubcategory
      tags: [categories]
      summary: Update a subcategory
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    delete:
      operationId: deleteSubcategory
      tags: [categories]
      summary: Delete a subcategory
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: Deleted

  # ── PRODUCTS ──────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/ref/{orderRef}:
    get:
      operationId: getOrderByRef
      tags: [orders]
      summary: Get an order by reference number (e.g. MH-000001)
      parameters:
        - { name: orderRef, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── SETTINGS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:          { type: number }
        name:        { type: string }
        slug:        { type: string }
        description: { type: ["string", "null"] }
        image:       { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:        { type: string }
        slug:        { type: string }
        description: { type: string }
        image:       { type: string }
    CategoryUpdate:
      type: object
      properties:
        name:        { type: string }
        slug:        { type: string }
        description: { type: string }
        image:       { type: string }
    Subcategory:
      type: object
      required: [id, categoryId, name, slug, createdAt]
      properties:
        id:         { type: number }
        categoryId: { type: number }
        name:       { type: string }
        slug:       { type: string }
        createdAt:  { type: string, format: date-time }
    SubcategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
    SubcategoryUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }

    # ── PRODUCT ─────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [productId, productName, productImage, quantity, unitPrice, totalPrice]
      properties:
        productId:    { type: number }
        productName:  { type: string }
        productImage: { type: string }
        quantity:     { type: number }
        unitPrice:    { type: number }
        totalPrice:   { type: number }
    OrderTimeline:
      type: object
      required: [status, timestamp]
      properties:
        status:    { type: string }
        timestamp: { type: string, format: date-time }
        note:      { type: ["string", "null"] }
    Order:
      type: object
      required: [id, orderRef, fullName, phone, state, city, address, items, subtotal, paymentMethod, paymentStatus, orderStatus, timeline, createdAt]
      properties:
        id:             { type: number }
        orderRef:       { type: string }
        fullName:       { type: string }
        phone:          { type: string }
        state:          { type: string }
        city:           { type: string }
        address:        { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:       { type: number }
        paymentMethod:  { type: string }
        paymentStatus:  { type: string }
        orderStatus:    { type: string }
        flutterwaveRef: { type: ["string", "null"] }
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/OrderTimeline"
        createdAt:      { type: string, format: date-time }
    OrderInput:
      type: object
      required: [fullName, phone, state, city, address, items, paymentMethod]
      properties:
        fullName:       { type: string }
        phone:          { type: string }
        state:          { type: string }
        city:           { type: string }
        address:        { type: string }
        paymentMethod:  { type: string }
        flutterwaveRef: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderStatusUpdate:
      type: object
      required: [orderStatus]
      properties:
        orderStatus:   { type: string }
        paymentStatus: { type: string }
        note:          { type: string }
...