# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderTimeline } from './orderTimeline';

export interface Order {
  id: number;
  orderRef: string;
  fullName: string;
  phone: string;
  state: string;
  city: string;
  address: string;
  items: OrderItem[];
  subtotal: number;
  paymentMethod: string;
  paymentStatus: string;
  orderStatus: string;
  /** @nullable */
  flutterwaveRef?: string | null;
  timeline: OrderTimeline[];
  createdAt: Date;
}